+++ volatilipy/helper_functions.py
@@ -98,34 +98,12 @@ def _generate_date_df(
     date_df = date_df.to_frame(index=True, name="expiryDate")
     date_df["QuantlibDate"] = date_df.apply(_datetime_to_quantlib_date, axis=1)
     date_df["tau"] = date_df.apply(
-        _calculate_tau, args=(vol_surface.referenceDate(), ql.ActualActual()), axis=1
+        _calculate_tau, args=(vol_surface.referenceDate(), ql.ActualActual(ql.ActualActual.ISDA)), axis=1
     )
     date_df["days_to_maturity"] = date_df.apply(
-        _calculate_dtm, args=(vol_surface.referenceDate(), ql.ActualActual()), axis=1
+        _calculate_dtm, args=(vol_surface.referenceDate(), ql.ActualActual(ql.ActualActual.ISDA)), axis=1
     )
     return date_df
-
-
-def _calculate_tau(
-    end_date: ql.Date,
-    start_date: ql.Date,
-    day_count_convention: ql.DayCounter,
-) -> float:
-    """Function to calculate a tau/difference in time from Quantlib objects
-
-    Args:
-        end_date (ql.Date): End date
-        start_date (ql.Date): Start date
-        day_count_convention (ql.DayCounter): What basis of days to use;
-            ActualActual, 30/360, etc.
-
-    Returns:
-        float: tau in years of the difference between the two dates
-    """
-
-    if isinstance(end_date, pd.core.series.Series):
-        end_date = end_date["QuantlibDate"]
-    return day_count_convention.yearFraction(start_date, end_date)
 
 
 def _calculate_dtm(
@@ -196,34 +174,6 @@ def _create_ql_vol_grid(
     return implied_vols
 
 
-# Helper functions for the implied vol object
-def _get_option_rfr(
-    risk_free_rates: pd.DataFrame,
-    position: namedtuple,
-    day_count_convention: ql.DayCounter,
-    rate_column_name: str,
-) -> float:
-    """Method to pull market data that is specific to one option (namely volatility
-    and a risk free rate)
-
-    Args:
-        risk_free_rates (pd.DataFrame): Dataframe containing risk free rates. Used to
-            pull the risk free rate corresponding to the date of maturity of the option
-        position (namedtuple): A named tuple that represents one option position; used to
-            retrieve the maturity date and strike
-        day_count_convention (QuantLibday_count_convention): Method to determine how to count days
-        between two dates
-    Returns:
-        risk_free_rate: The risk free rate
-    """
-
-    rate = ql.SimpleQuote(risk_free_rates[rate_column_name][position.exercise_date])
-    risk_free_rate = ql.FlatForward(
-        0, ql.TARGET(), ql.QuoteHandle(rate), day_count_convention
-    )
-    return (risk_free_rate,)
-
-
 def _setup_quantlib_economy(
     valuation_date: datetime, index_values: float64, dividend_yields: float64
 ) -> Tuple:
@@ -241,7 +191,7 @@ def _setup_quantlib_economy(
         valuation_date.day, valuation_date.month, valuation_date.year
     )
     ql.Settings.instance().evaluationDate = valuation_date_for_quantlib
-    day_count_convention = ql.ActualActual()
+    day_count_convention = ql.ActualActual(ql.ActualActual.ISDA)
 
     spot = index_values
     spot_quote = ql.QuoteHandle(ql.SimpleQuote(spot))
@@ -337,7 +287,7 @@ def _average_across_options(
         option_df.groupby(
             [column_name_mapping["strike"], column_name_mapping["expiration"]]
         )
-        .mean()
+        .mean(numeric_only=True)
         .reset_index()
     )
 
@@ -405,10 +355,8 @@ def _drop_valdate_row(option_df: pd.DataFrame, valuation_date: date) -> pd.DataF
         pd.Timestamp(valuation_date) in option_df.index
         or valuation_date in option_df.index
     ):
-        cleaned_df = option_df.drop([valuation_date])
-    else:
-        cleaned_df = option_df
-    return cleaned_df
+        option_df = option_df.drop([valuation_date])
+    return option_df
 
 
 def _sort_df(option_df: pd.DataFrame):
@@ -425,31 +373,23 @@ def _datetime_to_quantlib_date(
     Args:
         input_datetime (Union[datetime, pd.Series]): Either a datetime object or a pandas
             series object that contains datetime objects
+        datetime_column_name (str, Optional): the index value of the datetime to convert.
+            This is only used if input_datetime is a Series containing more than one row.
+            Optional, defaults to 'exercise_date'.
 
     Returns:
         ql.Date: The same dates, represented as Quantlib objects
     """
     if isinstance(input_datetime, pd.core.series.Series):
-        cols = input_datetime.shape[0]
-        if cols > 1:
-            return_date = ql.Date(
-                input_datetime.loc[datetime_column_name].day,
-                input_datetime.loc[datetime_column_name].month,
-                input_datetime.loc[datetime_column_name].year,
-            )
+        if len(input_datetime.index) > 1:
+            input_datetime = input_datetime.loc[datetime_column_name]
         else:
-            return_date = ql.Date(
-                input_datetime[0].day,
-                input_datetime[0].month,
-                input_datetime[0].year,
-            )
-    else:
-        return_date = ql.Date(
+            input_datetime = input_datetime[0]
+    return ql.Date(
             input_datetime.day,
             input_datetime.month,
             input_datetime.year,
         )
-    return return_date
 
 
 def _create_quantlib_option(dataframe_row: pd.Series) -> ql.EuropeanOption: